Remove pluggy and disable rope completions by default
